Update disability.rst
In the US, Emotional Disability is a big catch-all label from the education system that can include almost any "brain" issue (literally everything from bipolar disorder to Tourette's syndrome) and also the effects of trauma/systemic oppression. In practice, it is often used as a cipher for "student's behavior disrupts the classroom." This is generally a label put on people to assign them to a special ed class, not a label that individuals use for themselves.